Moved enrichers to new projects. Serilog.Enrichers.Environment contains UserName and Machine Name. Serilog.Erichers.Thread contains ThreadId. Serilog.Enrichers.Process contains ProcessId.
