+++ src/Serilog.Enrichers.Process/ProcessLoggerConfigurationExtensions.cs
@@ -1,0 +1,37 @@
+﻿// Copyright 2013-2016 Serilog Contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+using System;
+using System.Diagnostics;
+using Serilog.Configuration;
+using Serilog.Enrichers;
+
+namespace Serilog
+{
+    public static class ProcessLoggerConfigurationExtensions
+    { 
+        /// <summary>
+        /// Enrich log events with a ProcessId property containing the current <see cref="Process.Id"/>.
+        /// </summary>
+        /// <param name="enrichmentConfiguration">Logger enrichment configuration.</param>
+        /// <returns>Configuration object allowing method chaining.</returns>
+        public static LoggerConfiguration WithProcessId(
+           this LoggerEnrichmentConfiguration enrichmentConfiguration)
+        {
+            if (enrichmentConfiguration == null) throw new ArgumentNullException(nameof(enrichmentConfiguration));
+            return enrichmentConfiguration.With<ProcessIdEnricher>();
+        } 
+    }
+}
