+++ src/Serilog.Enrichers.Thread/project.json
@@ -1,0 +1,21 @@
+﻿{
+  "version": "2.0.0-beta-*",
+  "description": "The thread enricher for Serilog.",
+  "authors": [ "Serilog Contributors" ],
+  "tags": [ "serilog", "thread", "enricher" ],
+  "projectUrl": "http://serilog.net",
+  "licenseUrl": "http://www.apache.org/licenses/LICENSE-2.0",
+  "iconUrl": "http://serilog.net/images/serilog-extension-nuget.png",
+  "dependencies": {
+    "Serilog": { "target": "project" }
+  },
+  "compilationOptions": {
+    "keyFile": "../../assets/Serilog.snk"
+  },
+  "frameworks": {
+    "net45": {
+    },
+    "dotnet5.1": { 
+    }
+  }
+}
