+++ test/Serilog.Tests/app.config
@@ -4,6 +4,9 @@
     <add key="serilog:enrich:with-property:Path" value="%PATH%" />
     <add key="custom1:serilog:minimum-level" value="Warning" />
     <add key="custom2:serilog:minimum-level" value="Error" />
+    <add key="serilog:using:Environment" value="Serilog.Enrichers.Environment" />
+    <add key="serilog:using:Process" value="Serilog.Enrichers.Process" />
+    <add key="serilog:using:Thread" value="Serilog.Enrichers.Thread" />
     <add key="serilog:enrich:WithThreadId"/>
     <add key="serilog:enrich:WithMachineName"/>
     <add key="serilog:enrich:WithProcessId"/>
